requirements.txt
pillow
scikit-image
numpy